\name{ecworkset}
\alias{ecworkset}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Merge time series and returns the workset dataframe
}
\description{
Merge time series with common indexes (times) and returns the \emph{workset} data frame to be used as input for the data cleaning procedure via \code{\link{cleanFlux}} function.
}
\usage{
ecworkset(path_EPout, path_EPqc, path_EPmd, path_QCstat, path_output=NULL, FileName=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path_EPout}{
  path where the fulloutput file generated by EddyPro software is stored.
}
  \item{path_EPqc}{
path where the qcdetails file generated by EddyPro software is stored.
}
  \item{path_EPmd}{
path where the metadata file generated by EddyPro software is stored.
}
  \item{path_QCstat}{
path where the QC statistics file generated by \code{\link{qcStat}} is stored.
}
  \item{path_output}{
path where the results will be stored. Default is NULL.
}
  \item{FileName}{
a character string naming a file for writing. Default is NULL.
}
}
\details{
Returns a dataframe ordered by common indexes (times) containing a set of variables selected from the fulloutput, metadata, qcdetails files generated by EddyPro via \code{\link{eddypro_run}} and the statistics of the quality control routines (Vitale et al, 2019) generated by \code{\link{qcStat}}.
}
\value{
A dataframe object to be used as input of the \code{\link{cleanFlux}}. For the meaning of variables and units see the EddyPro manual and the \code{\link{qcStat}} function description.
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
LI-COR Biosciences: EddyPro 7.0.4: Help and User's Guide, LI-COR Biosciences, Lincoln, Nebraska USA, \url{www.licor.com/EddyPro}, 2019.

Vitale, D. Fratini, G. Bilancia, M. Nicolini, G. Sabbatini, S. Papale, D.
A robust data cleaning procedure for eddy covariance flux measurements,
Biogeosciences, 17, 1367--1391, doi: \url{https://doi.org/10.5194/bg-17-1367-2020}, 2020.
}
\author{
Domenico Vitale, Dario Papale
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
PATH_EPOUT <- system.file("extdata", "eddypro_DE-HoH_full_output_example.csv", package = "RFlux")
PATH_EPQC <- system.file("extdata", "eddypro_DE-HoH_qc_details_example.csv", package = "RFlux")
PATH_EPMD <- system.file("extdata", "eddypro_DE-HoH_metadata_example.csv", package = "RFlux")
PATH_QCSTAT <- system.file("extdata", "qcStat_example.csv", package = "RFlux")

WorkSet <- ecworkset(path_EPout=PATH_EPOUT,
                     path_EPqc=PATH_EPQC,
                     path_EPmd=PATH_EPMD,
                     path_QCstat=PATH_QCSTAT,
                     path_output=NULL,FileName=NULL)
str(WorkSet)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datagen}% __ONLY ONE__ keyword per line
